+++ package.json
@@ -63,11 +63,12 @@
     "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
     "rc": "^1.2.8",
-    "react": "^16.4.2",
+    "react": "^16.8.0",
     "react-contextmenu": "^2.9.4",
-    "react-dom": "^16.4.2",
+    "react-dom": "^16.8.0",
     "react-intl": "^2.4.0",
     "react-qr-svg": "^2.1.0",
+    "react-window": "^1.8.1",
     "rimraf": "^2.6.2",
     "run-parallel": "^1.1.9",
     "run-series": "^1.1.8",
